# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Education CRM Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get list of courses",
                "description": "Returns list of courses",
                "operationId": "df5cfe3ec2c1d6bdc815652059af1502",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Math 101"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Basic mathematics"
                                            },
                                            "price": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "room_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Create a new course",
                "operationId": "e3735861009fb8f5bb551fcd23c4e6aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Math 101"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Basic mathematics"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "room_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "course": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Math 101"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Basic mathematics"
                                                },
                                                "price": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "room_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/{courseId}/add-student": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Add a student to a course",
                "description": "Adds a student to a specific course",
                "operationId": "52b4b844cd15353981f6ae8fa87732d5",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "ID of the course",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "student_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Student added to course successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Student added to course successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid data provided"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Student not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get a course by ID",
                "operationId": "9d444899c0eea57b198b07745dfeb99b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the course to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "course": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Math 101"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Basic mathematics"
                                                },
                                                "price": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "student-count": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "room_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Courses"
                ],
                "summary": "Update a course",
                "operationId": "acc1fb25618eca0ca4f7c9351f89159e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of course to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Math 101"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Basic mathematics"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "room_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "course": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Math 101"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Basic mathematics"
                                                },
                                                "price": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "room_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Courses"
                ],
                "summary": "Delete a course",
                "operationId": "96ec6735185fdd95188d94afd17c82ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of course to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Course deleted successfully"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}